requirements-dev.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements-dev.txt requirements-dev.in
#
ansible==4.4.0
    # via -r requirements-dev.in
ansible-core==2.11.3
    # via ansible
ansible-lint==5.1.2
    # via
    #   -r requirements-dev.in
    #   molecule
appdirs==1.4.4
    # via black
arrow==0.15.8
    # via jinja2-time
attrs==19.3.0
    # via pytest
bcrypt==3.1.7
    # via paramiko
binaryornot==0.4.4
    # via cookiecutter
black==21.7b0
    # via -r requirements-dev.in
bracex==2.1.1
    # via wcmatch
cerberus==1.3.2
    # via molecule
certifi==2020.6.20
    # via requests
cffi==1.14.1
    # via
    #   bcrypt
    #   cryptography
    #   pynacl
chardet==3.0.4
    # via
    #   binaryornot
    #   requests
click==8.0.1
    # via
    #   black
    #   click-help-colors
    #   cookiecutter
    #   molecule
click-help-colors==0.9
    # via molecule
colorama==0.4.3
    # via rich
commonmark==0.9.1
    # via rich
cookiecutter==1.7.3
    # via molecule
cryptography==3.3.2
    # via
    #   ansible-core
    #   paramiko
distro==1.5.0
    # via selinux
docker==4.4.3
    # via molecule-docker
enrich==1.2.6
    # via
    #   ansible-lint
    #   molecule
flake8==3.9.2
    # via
    #   -r requirements-dev.in
    #   flake8-docstrings
    #   flake8-isort
flake8-docstrings==1.6.0
    # via -r requirements-dev.in
flake8-isort==4.0.0
    # via -r requirements-dev.in
idna==2.10
    # via requests
iniconfig==1.0.0
    # via pytest
isort==5.9.3
    # via flake8-isort
jinja2==2.11.3
    # via
    #   ansible-core
    #   cookiecutter
    #   jinja2-time
    #   molecule
jinja2-time==0.2.0
    # via cookiecutter
markupsafe==1.1.1
    # via jinja2
mccabe==0.6.1
    # via flake8
molecule==3.4.0
    # via
    #   -r requirements-dev.in
    #   molecule-docker
molecule-docker==0.2.4
    # via -r requirements-dev.in
more-itertools==8.4.0
    # via pytest
mypy-extensions==0.4.3
    # via black
packaging==20.4
    # via
    #   ansible-core
    #   ansible-lint
    #   molecule
    #   pytest
paramiko==2.7.1
    # via molecule
pathspec==0.9.0
    # via black
pluggy==0.13.1
    # via
    #   molecule
    #   pytest
poyo==0.5.0
    # via cookiecutter
py==1.10.0
    # via pytest
pycodestyle==2.7.0
    # via flake8
pycparser==2.20
    # via cffi
pydocstyle==6.1.1
    # via flake8-docstrings
pyflakes==2.3.1
    # via flake8
pygments==2.8.0
    # via rich
pynacl==1.4.0
    # via paramiko
pyparsing==2.4.7
    # via packaging
pytest==6.0.1
    # via pytest-testinfra
pytest-testinfra==6.4.0
    # via -r requirements-dev.in
python-dateutil==2.8.1
    # via arrow
python-slugify==4.0.1
    # via cookiecutter
pyyaml==5.4
    # via
    #   ansible-core
    #   ansible-lint
    #   molecule
regex==2021.8.3
    # via black
requests==2.25.1
    # via
    #   cookiecutter
    #   docker
resolvelib==0.5.4
    # via ansible-core
rich==9.11.1
    # via
    #   ansible-lint
    #   enrich
    #   molecule
ruamel.yaml==0.16.10
    # via ansible-lint
ruamel.yaml.clib==0.2.6
    # via ruamel.yaml
selinux==0.2.1
    # via
    #   molecule
    #   molecule-docker
six==1.15.0
    # via
    #   bcrypt
    #   cookiecutter
    #   cryptography
    #   docker
    #   packaging
    #   pynacl
    #   python-dateutil
    #   tenacity
    #   websocket-client
snowballstemmer==2.1.0
    # via pydocstyle
subprocess-tee==0.3.2
    # via molecule
tenacity==7.0.0
    # via ansible-lint
testfixtures==6.18.1
    # via flake8-isort
text-unidecode==1.3
    # via python-slugify
toml==0.10.1
    # via pytest
tomli==1.2.1
    # via black
typing-extensions==3.7.4.3
    # via rich
urllib3==1.26.5
    # via requests
wcmatch==8.1.1
    # via ansible-lint
websocket-client==0.57.0
    # via docker

# The following packages are considered to be unsafe in a requirements file:
# setuptools
